[asan] Reduce flakiness in stack-overflow detection
IsStackOverflow only treats accesses within 512 bytes of SP as stack-overflow. This should really be the size of a page instead.

The scariness_score_test.cc triggers stack overflow with frames that are even larger than a page, which can also trigger a fault that will not be recognized as stack-overflow. Let's just use smaller frames.



git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@329980 91177308-0d34-0410-b5e6-96231b3b80d8