[CMake] Fix Libc++ Modules build.
When building the dylib, the C++ headers are fundamentally non-module.
They require special versions of the headers in order to provide C++03 and
legacy ABI definitions. This causes ODR issues when modules are enabled
during both the build and the usage of the libc++ headers.

This patch fixes the build error by disabling modules when building the
libc++ sources.


git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@334220 91177308-0d34-0410-b5e6-96231b3b80d8